+++ spec/configuration_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe DatabaseConsistency::Configuration do
   shared_examples 'checker' do |value|
     context 'checker' do
       specify do
-        expect(configuration.enabled?('ColumnPresenceChecker')).to eq(value)
+        expect(configuration.enabled?('User', 'email', 'ColumnPresenceChecker')).to eq(value)
       end
     end
   end
@@ -14,7 +14,7 @@ RSpec.describe DatabaseConsistency::Configuration do
   shared_examples 'model' do |value|
     context 'model' do
       specify do
-        expect(configuration.enabled?('ColumnPresenceChecker', 'User')).to eq(value)
+        expect(configuration.enabled?('User')).to eq(value)
       end
     end
   end
@@ -22,7 +22,7 @@ RSpec.describe DatabaseConsistency::Configuration do
   shared_examples 'key' do |value|
     context 'key' do
       specify do
-        expect(configuration.enabled?('ColumnPresenceChecker', 'User', 'email')).to eq(value)
+        expect(configuration.enabled?('User', 'email')).to eq(value)
       end
     end
   end
@@ -30,32 +30,32 @@ RSpec.describe DatabaseConsistency::Configuration do
   context 'when checker is disabled' do
     let(:file_path) { 'checker_disabled.yml' }
 
+    include_examples 'model', true
+    include_examples 'key', true
     include_examples 'checker', false
-    include_examples 'model', false
-    include_examples 'key', false
   end
 
   context 'when model is disabled' do
     let(:file_path) { 'model_disabled.yml' }
 
-    include_examples 'checker', true
     include_examples 'model', false
     include_examples 'key', false
+    include_examples 'checker', false
   end
 
   context 'when key is disabled' do
     let(:file_path) { 'key_disabled.yml' }
 
-    include_examples 'checker', true
     include_examples 'model', true
     include_examples 'key', false
+    include_examples 'checker', false
 
     context 'when compact' do
-      let(:file_path) { 'compact_key_disabled.yml' }
+      let(:file_path) { 'compact_checker_disabled.yml' }
 
-      include_examples 'checker', true
       include_examples 'model', true
-      include_examples 'key', false
+      include_examples 'key', true
+      include_examples 'checker', false
     end
   end
 end